+++ src/pages/card_pages/card_page_filter.py
@@ -20,8 +20,7 @@ layout = dbc.Row([
         dbc.InputGroup(
             [
                 dbc.InputGroupText('Account'),
-                dcc.Dropdown(store_util.get_account_names(),
-                             id=card_page_ids.dropdown_user_selection,
+                dcc.Dropdown(id=card_page_ids.dropdown_user_selection,
                              className='dbc',
                              style={'width': '70%'},
                              ),
@@ -46,13 +45,14 @@ layout = dbc.Row([
 
 
 @app.callback(Output(card_page_ids.dropdown_user_selection, 'value'),
+              Output(card_page_ids.dropdown_user_selection, 'options'),
               Input(card_page_ids.dropdown_user_selection, 'state'),
               Input(nav_ids.trigger_daily, 'data'))
 def update_account_value(state, daily_trigger):
     if load_with_account_name:
-        return load_with_account_name
+        return load_with_account_name, store_util.get_account_names()
     else:
-        return store_util.get_first_account_name()
+        return store_util.get_first_account_name(), store_util.get_account_names()
 
 
 @app.callback(Output(card_page_ids.filtered_cards_top_df, 'data'),
@@ -115,18 +115,16 @@ def update_card_list(user_selection, tigger):
     else:
         df = analyse.filter_battles(store.battle_big, filter_account=user_selection)
 
+    value = None
+    played_cards = []
     if not df.empty:
         played_cards = df.card_name.unique().tolist()
 
         if load_with_card_id:
-            value = df.loc[df.card_detail_id == int(load_with_card_id)].card_name.unique().tolist()[0]
+            df = df.loc[df.card_detail_id == int(load_with_card_id)]
+            if not df.empty:
+                value = df.card_name.unique().tolist()[0]
         else:
             if len(played_cards) > 0:
                 value = played_cards[0]
-            else:
-                value = None
-    else:
-        value = None
-        played_cards = []
     return value, played_cards
-