+++ src/pages/config_pages/config_page_ids.py
@@ -1,0 +1,7 @@
+accounts_updated = 'accounts_updated'
+add_button = 'add-account-btn'
+remove_button = 'remove-account-btn'
+error_text_account = 'error-text-account'
+current_accounts = 'current-accounts'
+account_name_input = 'account-name-input'
+button_div = 'button-div'