+++ src/pages/navigation_pages/nav_ids.py
@@ -1,1 +1,5 @@
 trigger_daily = 'trigger-daily-update'
+load_new_values = 'load-new-values'
+progress_daily = 'progress-daily'
+progress_season = 'progress-season'
+interval_global = 'interval-global'