+++ src/pages/portfolio_pages/portfolio_deposit.py
@@ -5,7 +5,22 @@ from main import app
 import dash_bootstrap_components as dbc
 from dash import html, dcc, Output, Input, ctx
 
+from src.configuration import config
+from src.pages.navigation_pages import nav_ids
 from src.utils import store_util, portfolio_util
+
+
+def get_div_style():
+    if config.read_only:
+        return {'display': 'none'}
+    else:
+        return {'display': 'block'}
+
+
+def get_readonly_text():
+    if config.read_only:
+        return html.H4("Read only mode not possible to deposit/withdraw investment", className='text-warning')
+    return ""
 
 
 def get_deposit_layout():
@@ -13,58 +28,69 @@ def get_deposit_layout():
         dbc.AccordionItem(
             [
                 html.P('This deposit or removes an investment on a certain date'),
-                dbc.InputGroup(
-                    [
-                        dbc.InputGroupText('Date'),
-                        dcc.DatePickerSingle(
-                            id='my-date-picker-single',
-                            min_date_allowed=date(2015, 8, 5),
-                            initial_visible_month=date.today(),
-                            date=date.today(),
-                            className='dbc',
-                        ),
-                    ], className='mb-3'),
-                dbc.InputGroup(
-                    [
-                        dbc.InputGroupText('Account'),
-                        dcc.Dropdown(options=store_util.get_account_names(),
-                                     value=store_util.get_first_account_name(),
-                                     id='dropdown-user-selection',
-                                     style={'min-width': '300px'},
-                                     className='dbc'),
-                    ], className='mb-3', ),
-                dbc.InputGroup(
-                    [
-                        dbc.InputGroupText('Amount'),
-                        dbc.Input(id='amount', type='number', pattern='[0-9]'),
-                    ], className='mb-3'),
-                dbc.Row(
-                    [
-                        dbc.Col(dbc.Button('Deposit', id='deposit', className='ml-auto'),
-                                width=2,
-                                className='mb-3'),
-                        dbc.Col(dbc.Button('Withdraw', id='withdraw', className='ml-auto'),
-                                width=2,
-                                className='mb-3')
-                    ]),
-                html.Div([
-                    html.P(id='error-text')
-                ])
+                html.Div(style=get_div_style(),
+                    children=[
+                        dbc.InputGroup(
+                        [
+                            dbc.InputGroupText('Date'),
+                            dcc.DatePickerSingle(
+                                id='my-date-picker-single',
+                                min_date_allowed=date(2015, 8, 5),
+                                initial_visible_month=date.today(),
+                                date=date.today(),
+                                className='dbc',
+                            ),
+                        ], className='mb-3'),
+                    dbc.InputGroup(
+                        [
+                            dbc.InputGroupText('Account'),
+                            dcc.Dropdown(id='dropdown-user-selection-deposit',
+                                         style={'min-width': '300px'},
+                                         className='dbc'),
+                        ], className='mb-3', ),
+                    dbc.InputGroup(
+                        [
+                            dbc.InputGroupText('Amount'),
+                            dbc.Input(id='amount', type='number', pattern='[0-9]'),
+                        ], className='mb-3'),
+                    dbc.Row(
+                        [
+                            dbc.Col(dbc.Button('Deposit', id='deposit', className='ml-auto'),
+                                    width=2,
+                                    className='mb-3'),
+                            dbc.Col(dbc.Button('Withdraw', id='withdraw', className='ml-auto'),
+                                    width=2,
+                                    className='mb-3')
+                        ]),
+                    html.Div([
+                        html.P(id='error-text')
+                    ])
+                ]),
+                html.Div(children=get_readonly_text()),
             ],
             title='deposit/withdraw investment',
         ),
         start_collapsed=True,
     ),
 
+
+@app.callback(Output('dropdown-user-selection-deposit', 'value'),
+              Output('dropdown-user-selection-deposit', 'options'),
+              Input(nav_ids.trigger_daily, 'data'),
+              )
+def update_user_list(tigger):
+    return store_util.get_first_account_name(), store_util.get_account_names()
+
+
 @app.callback(
     Output('trigger-portfolio-update', 'data'),
     [Input('deposit', 'n_clicks'),
-     Input('dropdown-user-selection', 'value'),
+     Input('dropdown-user-selection-deposit', 'value'),
      Input('my-date-picker-single', 'date'),
      Input('amount', 'value')],
 )
 def deposit_action(deposit_clicks, account, my_date, amount):
-    if ctx.triggered_id == 'deposit':
+    if ctx.triggered_id == 'deposit' and account:
         portfolio_util.update_investment(account, amount, my_date)
         logging.info('Deposit: ' + str(my_date) + ' amount: ' + str(amount))
         return True
@@ -74,12 +100,12 @@ def deposit_action(deposit_clicks, account, my_date, amount):
 @app.callback(
     Output('trigger-portfolio-update', 'data'),
     [Input('withdraw', 'n_clicks'),
-     Input('dropdown-user-selection', 'value'),
+     Input('dropdown-user-selection-deposit', 'value'),
      Input('my-date-picker-single', 'date'),
      Input('amount', 'value')],
 )
 def withdraw_action(withdraw_clicks, account, my_date, amount):
-    if ctx.triggered_id == 'withdraw':
+    if ctx.triggered_id == 'withdraw' and account:
         portfolio_util.update_investment(account, amount * -1, my_date)
         logging.info('Withdraw: ' + str(my_date) + ' amount: ' + str(amount))
         return True