+++ src/utils/store_util.py
@@ -3,8 +3,10 @@ import os
 
 import pandas as pd
 
+from src import portfolio, collection_store, battle_store
 from src.api import spl
 from src.configuration import store, config
+from src.utils import progress_util
 
 
 def update_season_end_dates():
@@ -151,3 +153,17 @@ def get_rule_sets_list():
 
 def get_last_season_values(df, users, season_id_column='season_id'):
     return df.loc[(df.player.isin(users)) & (df[season_id_column] == df[season_id_column].max())].copy()
+
+
+def update_data():
+    progress_util.set_daily_title('Update collection')
+    collection_store.update_collection()
+
+    progress_util.set_daily_title('Update battles')
+    battle_store.process_battles()
+
+    progress_util.set_daily_title('Update portfolio')
+    portfolio.update_portfolios()
+
+    save_stores()
+    progress_util.update_daily_msg('Done')